chore(id): try again with this fix
affects: @id/tenant
