+++ applications/id/tenant/src/App.tsx
@@ -4,13 +4,11 @@ import {
   Loader,
   ToastProvider,
   Typography,
-  Window,
 } from '@motech-development/breeze-ui';
 import React, { FC, lazy, memo, Suspense } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Route, Router, Switch } from 'react-router-dom';
 import history from './history';
-import AppTitle from './components/AppTitle';
 import Wrapper from './components/Wrapper';
 
 const Login = lazy(() => import('./pages/Login'));
@@ -18,7 +16,6 @@ const Reset = lazy(() => import('./pages/Reset'));
 
 const App: FC = () => {
   const { t } = useTranslation();
-  const showContent = window.passwordReset || window.config;
 
   return (
     <Wrapper>
@@ -30,20 +27,12 @@ const App: FC = () => {
 
       <Suspense fallback={<Loader />}>
         <ToastProvider>
-          {showContent ? (
-            <Window>
-              <AppTitle />
-
-              <Router history={history}>
-                <Switch>
-                  <Route exact path="/login" component={Login} />
-                  <Route exact path="/lo/reset" component={Reset} />
-                </Switch>
-              </Router>
-            </Window>
-          ) : (
-            <Loader />
-          )}
+          <Router history={history}>
+            <Switch>
+              <Route exact path="/login" component={Login} />
+              <Route exact path="/lo/reset" component={Reset} />
+            </Switch>
+          </Router>
         </ToastProvider>
       </Suspense>
