+++ applications/id/tenant/src/__tests__/App.tsx
@@ -13,51 +13,4 @@ describe('App', () => {
 
     await expect(findByText('app-name')).resolves.toBeInTheDocument();
   });
-
-  it('should show loading spinner if configs is not set', () => {
-    const { container } = render(
-      <TestProvider>
-        <App />
-      </TestProvider>,
-    );
-    const loader = container.querySelector('circle');
-
-    expect(loader).toBeInTheDocument();
-  });
-
-  it('should not show loading spinner if log in config is set', () => {
-    window.config = {
-      auth0Domain: 'https://auth0.com',
-      callbackURL: 'https://my-app.com',
-      clientID: 'client-id',
-      internalOptions: {},
-    };
-
-    const { container } = render(
-      <TestProvider>
-        <App />
-      </TestProvider>,
-    );
-    const loader = container.querySelector('circle');
-
-    expect(loader).not.toBeInTheDocument();
-  });
-
-  it('should not show loading spinner if reset password config is set', () => {
-    window.passwordReset = {
-      csrfToken: 'token',
-      email: 'text@example.com',
-      passwordPolicy: 'good',
-      ticket: 'ticket',
-    };
-
-    const { container } = render(
-      <TestProvider>
-        <App />
-      </TestProvider>,
-    );
-    const loader = container.querySelector('circle');
-
-    expect(loader).not.toBeInTheDocument();
-  });
 });
